unified-sort/requirements.txt
# Core
streamlit>=1.25
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
pillow>=9.5

# Optional
pillow-heif>=0.13
rawpy>=0.18
imageio>=2.31
send2trash>=1.8
